refactor: remove PaymentHandler and related payment handler classes refactor: update StagHerdServiceProvider to register new payment services and MP provider feat: implement ProviderRegistry for managing payment providers test: add CreatePaymentTest for cash payment scenarios

+++ src/Data/PaymentResultData.php
@@ -57,7 +57,7 @@ final readonly class PaymentResultData
         string $method,
         ?string $providerStatus = null,
         ?ProviderReferencesData $references = null,
-        ?float $amount = null,
+        ?int $amount = null,
         ?string $currency = null,
         ?NextActionData $nextAction = null,
         ?string $reason = null,
